Fix compiler warnings.

+++ dynamic_cloth/EventHandler.cpp
@@ -12,14 +12,14 @@ void EventHandler::check_events(sf::RenderWindow& app, ClothSimulator& cs) {
         if (event.type == sf::Event::Resized)
         {
             // update the view to the new size of the window
-            sf::FloatRect visibleArea(0, 0, event.size.width, event.size.height);
+            sf::FloatRect visibleArea(0.f, 0.f, (float)event.size.width, (float)event.size.height);
             app.setView(sf::View(visibleArea));
         }
 
         // Remove a point if we click near it.
         if (sf::Mouse::isButtonPressed(sf::Mouse::Left)) {
             sf::Vector2i localPosition = sf::Mouse::getPosition(app);
-            for (int ii = 0; ii < cs.cloth.points.size(); ii++) {
+            for (size_t ii = 0; ii < cs.cloth.points.size(); ii++) {
                 Particle& p = cs.cloth.points.at(ii);
                 if (std::abs(localPosition.x - p.x) < 5 && std::abs(localPosition.y - p.y) < 5) {
                     cs.cloth.remove_particle(ii);
