+++ dynamic_cloth/physics/ClothLoader.cpp
@@ -50,7 +50,7 @@ void ClothLoader::load_cloth_coordinates(Cloth &c, std::string& filepath) {
     }
     
     // Store the indicies and randomise them, allows us to iterate through array in random order.
-    for (int i = 0; i < c.points.size(); i++)
+    for (size_t i = 0; i < c.points.size(); i++)
         c.point_idicies.push_back(i);
     std::random_shuffle(c.point_idicies.begin(), c.point_idicies.end());
 }
